flask version/requirements.txt
langchain==0.3.26
flask==3.1.1 
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-google-genai==2.1.9
langchain-community==0.3.26
langchain-huggingface==0.3.1
tensorflow==2.13.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
werkzeug==3.0.1
h5py==3.10.0
streamlit==1.28.1
plotly==5.17.0
pandas==2.0.3